+++ apps/blog-app/src/pages/_document.tsx
@@ -1,32 +1,29 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document';
+import { Html, Head, Main, NextScript } from 'next/document';
 
-class MyDocument extends Document {
-  render() {
-    return (
-      <Html>
-        <Head>
-          {/*
+export default function Document() {
+  return (
+    <Html>
+      <Head>
+        {/*
              Preconnect might be part of nextjs font optimization by default in
              a later version, so this line can be dropped
              @link https://github.com/vercel/next.js/discussions/24645
           */}
-          <link
-            rel="preconnect"
-            href="https://fonts.gstatic.com"
-            crossOrigin="anonymous"
-          />
-          <link
-            href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700&display=swap"
-            rel="stylesheet"
-          />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
+        <link
+          rel="preconnect"
+          href="https://fonts.gstatic.com"
+          crossOrigin="anonymous"
+        />
+        <link
+          href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700&display=swap"
+          rel="stylesheet"
+        />
+      </Head>
 
-export default MyDocument;
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  );
+}