+++ apps/blog-app/src/types.d/react-svgr.d.ts
@@ -19,6 +19,6 @@
 
 declare module '*.svg' {
   import type React from 'react';
-  const SVG: React.VFC<React.SVGProps<SVGSVGElement>>;
-  export default SVG;
+  const svg: React.VFC<React.SVGProps<SVGSVGElement>>;
+  export default svg;
 }