+++ apps/remix-app/test/correct.test.tsx
@@ -1,0 +1,10 @@
+import { default as NiceWork } from '../app/routes/demos/correct';
+import { render, screen } from '@/test-utils';
+
+describe('Correct component', () => {
+  it('renders without crashing', () => {
+    render(<NiceWork />);
+    const appContent = screen.getByTestId('nice-work');
+    expect(appContent).toHaveTextContent('You got it right!');
+  });
+});