+++ apps/web-app/src/themes/tailwind/tailwind.theme.js
@@ -1,7 +1,7 @@
 // @ts-check
 // Keep this file as '.js' as it's included in tailwind.config.js
 
-const { browserFonts } = require('../global/browser-fonts');
+const { browserFonts } = require('../shared/browser-fonts');
 
 module.exports = {
   fontFamily: {