+++ lint-staged.config.base.js
@@ -1,20 +1,28 @@
+// @ts-check
+
 /**
  * lint-staged/prettier has an issue with special characters in filenames,
  * like the ones uses for nextjs dynamic routes (ie: [id].tsx...)
  * @link https://github.com/okonet/lint-staged/issues/676
  */
 const escape = require('shell-quote').quote;
-const { CLIEngine } = require('eslint');
+const { ESLint } = require('eslint');
 
-const cli = new CLIEngine({});
+const cli = new ESLint();
 const isWin = process.platform === 'win32';
 
+/**
+ * @param {string[]} filenames
+ */
 const escapeFileNamesForPrettier = (filenames) =>
   filenames
     .map((filename) => `"${isWin ? filename : escape([filename])}"`)
     .join(' ');
 
-module.exports = {
+/**
+ * @type {Record<string, (filenames: string[]) => string[]>}
+ */
+const rules = {
   '**/*.{js,jsx,ts,tsx}': (filenames) => {
     return [
       // react-hooks/exhaustive-deps must be kept off, a change made here can
@@ -29,3 +37,5 @@ module.exports = {
     return [`prettier --write ${escapeFileNamesForPrettier(filenames)}`];
   },
 };
+
+module.exports = rules;