+++ packages/ui-lib/jest.config.js
@@ -1,6 +1,6 @@
 // @ts-check
 const { defaults: tsjPreset } = require('ts-jest/presets');
-const { pathsToModuleNameMapper } = require('ts-jest/utils');
+const { pathsToModuleNameMapper } = require('ts-jest');
 
 const { getJestCachePath } = require('../../cache.config');