+++ packages/core-lib/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @your-org/core-lib
 
+## 3.9.0
+
+### Minor Changes
+
+- [#963](https://github.com/belgattitude/nextjs-monorepo-example/pull/963) [`5e25fab2`](https://github.com/belgattitude/nextjs-monorepo-example/commit/5e25fab2f0d620e999f536a1fb8e0ef45d56fd64) Thanks [@belgattitude](https://github.com/belgattitude)! - Upgraded to react 18.0.0-rc.0
+
 ## 3.8.1
 
 ### Patch Changes